♻️ Update react-email
