+++ packages/emails/src/templates/_components/notification-email.tsx
