--- packages/emails/src/templates/components/email-context.tsx
@@ -1,5 +1,0 @@
-export type EmailContext = {
-  logoUrl: string;
-  baseUrl: string;
-  domain: string;
-};
