+++ packages/emails/src/templates/finalized-host.tsx
@@ -1,8 +1,8 @@
 import { Column, Row, Section } from "@react-email/components";
 
-import { EmailContext } from "./components/email-context";
-import { EmailLayout } from "./components/email-layout";
-import { borderColor, Button, Text } from "./components/styled-components";
+import { defaultEmailContext, EmailContext } from "./_components/email-context";
+import { EmailLayout } from "./_components/email-layout";
+import { borderColor, Button, Text } from "./_components/styled-components";
 
 export interface FinalizeHostEmailProps {
   date: string;
@@ -25,7 +25,7 @@ export const FinalizeHostEmail = ({
   dow = "Fri",
   date = "Friday, 12th June 2020",
   time = "6:00 PM to 11:00 PM BST",
-  ctx,
+  ctx = defaultEmailContext,
 }: FinalizeHostEmailProps) => {
   return (
     <EmailLayout ctx={ctx} recipientName={name} preview="Final date booked!">
