+++ packages/emails/src/templates/finalized-participant.tsx
@@ -1,8 +1,8 @@
 import { Column, Row, Section } from "@react-email/components";
 
-import { EmailContext } from "./components/email-context";
-import { EmailLayout } from "./components/email-layout";
-import { borderColor, Button, Text } from "./components/styled-components";
+import { defaultEmailContext, EmailContext } from "./_components/email-context";
+import { EmailLayout } from "./_components/email-layout";
+import { borderColor, Button, Text } from "./_components/styled-components";
 
 export interface FinalizeParticipantEmailProps {
   date: string;
@@ -27,7 +27,7 @@ export const FinalizeParticipantEmail = ({
   dow = "Fri",
   date = "Friday, 12th June 2020",
   time = "6:00 PM to 11:00 PM BST",
-  ctx,
+  ctx = defaultEmailContext,
 }: FinalizeParticipantEmailProps) => {
   return (
     <EmailLayout ctx={ctx} recipientName={name} preview="Final date booked!">
