+++ packages/emails/src/templates/static/logo.png
